refactor: add plugin within nuxt core
